refactor the utils to utils and operations
