+++ exercises/image_gradient.hpp
@@ -16,9 +16,8 @@ private:
     std::map<std::tuple<size_t, size_t>, Eigen::MatrixXd> g_mags;
     std::map<std::tuple<size_t, size_t>, Eigen::MatrixXd> g_dirs;
     std::vector<std::vector<Eigen::MatrixXd>> blurred;
-    Eigen::MatrixXd sobel_kernel();
-    Eigen::MatrixXd sobel_x_kernel;
-    Eigen::MatrixXd sobel_y_kernel;
+    Eigen::MatrixXd sobel_kernel_x;
+    Eigen::MatrixXd sobel_kernel_y;
     std::tuple<size_t, size_t> make_key(size_t octave, size_t scale);
     void calculate_grads(size_t octave, size_t scale);
 public:
