+++ exercises/utils/counting_iterator.hpp
