+++ exercises/utils/cuda_types.cpp
@@ -1,0 +1,1 @@
+#include "cuda_types.hpp"
