+++ exercises/utils/cuda_types.cuh
@@ -1,0 +1,1 @@
+#pragma once
