+++ exercises/utils/operations.cu
@@ -1,0 +1,1 @@
+#include "operations.cuh"
