+++ exercises/utils/operations.cuh
@@ -1,0 +1,1 @@
+#pragma once
