+++ exercises/utils/utils.cuh
